requirements.txt
fastapi
uvicorn
sentence-transformers
pinecone-client
pydantic
